+++ http_status/test/http_status_test.exs
@@ -1,4 +1,0 @@
-defmodule HTTPStatusTest do
-  use ExUnit.Case
-  doctest HTTPStatus
-end